requirements.txt
openshift
pyjwt
psycopg2-binary
uvicorn
fastapi
daiquiri
passgen
passlib
bcrypt
python-multipart
python-dotenv
requests
authlib
six
flask